audiobook page  is  workinf well

+++ lib/views/dialogs.dart
@@ -4,13 +4,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:key_board_app/cubits/convert_and_reading/convert_and_reading_cubit.dart';
 import 'package:key_board_app/cubits/for_take_image/take_image_cubit.dart';
+import 'package:key_board_app/cubits/part_saved_books/part_audio_books_dart_cubit.dart';
+import 'package:key_board_app/models/book_on_audio.dart';
 import '../cubits/for_read_audio_book/reading_audio_book_cubit.dart';
 import '../cubits/saved_book/saved_books_cubit.dart';
 import '../models/audio_model.dart';
 import '../navigators/goto.dart';
 import '../pages/home_page.dart';
 import '../services/hive_service.dart';
-saveAudioDialog(BuildContext context, List<AudioModel> audioModel,
+
+saveAudioDialog(BuildContext context, List<AudioModel> audioBook,
     {bool isBack = false}) async {
   showDialog(
       context: context,
@@ -39,30 +42,38 @@ saveAudioDialog(BuildContext context, List<AudioModel> audioModel,
             TextButton(
                 onPressed: () async {
                   List<dynamic>? listOfAudio =
-                  await HiveDB.loadCountryCode(key: "listOfAudio");
+                      await HiveDB.loadCountryCode(key: "listOfAudio");
 
-                  for(int i=0; i<audioModel.length; i++){
-                    print(audioModel[i]);
-                    if (listOfAudio == null) {
-                      List<dynamic> list = [audioModel[i].toJson()];
-                      await HiveDB.saveData("listOfAudio", list);
-                      listOfAudio  = await HiveDB.loadCountryCode(key: "listOfAudio");
-                    } else {
-                      print(audioModel[i].toJson());
-                      listOfAudio.add(audioModel[i].toJson());
-                    }
+                  listOfAudio ??= [];
+
+                  String? langCode = HiveDB.loadLangCode()!;
+                  String part = langCode == "uz"
+                      ? "qism"
+                      : langCode == "ru"
+                          ? "часть"
+                          : "part";
+
+                  AudioBook book =
+                      AudioBook(name: audioBook[0].name, partOfBook: []);
+
+                  for (int i = 0; i < audioBook.length; i++) {
+                    AudioModel newAudio = AudioModel(
+                        name: (i + 1).toString() + " -$part",
+                        path: audioBook[i].path);
+
+                    book.partOfBook!.add(newAudio);
                   }
-                   print(listOfAudio);
+
+                  listOfAudio.add(book.toJson());
 
                   await HiveDB.saveData("listOfAudio", listOfAudio);
 
-
-
-
-
                   GOTO.pop(context);
                 },
-                child:  Text("save",style: TextStyle(color:Colors.green.shade700,fontSize: 15)).tr()),
+                child: Text("save",
+                        style: TextStyle(
+                            color: Colors.green.shade700, fontSize: 15))
+                    .tr()),
             TextButton(
                 onPressed: () async {
                   if (isBack) {
@@ -71,7 +82,10 @@ saveAudioDialog(BuildContext context, List<AudioModel> audioModel,
                     GOTO.pop(context);
                   }
                 },
-                child:  Text("no",style: TextStyle(color:Colors.green.shade700,fontSize: 15)).tr()),
+                child: Text("no",
+                        style: TextStyle(
+                            color: Colors.green.shade700, fontSize: 15))
+                    .tr()),
           ],
         );
       });
@@ -105,43 +119,59 @@ errorDialog(BuildContext context, bool isPage) {
                 onPressed: () {
                   GOTO.pushRpUntil(context, HomePage());
                 },
-                child: Text("go_home",style: TextStyle(color:Colors.green.shade700,fontSize: 15)).tr()),
+                child: Text("go_home",
+                        style: TextStyle(
+                            color: Colors.green.shade700, fontSize: 15))
+                    .tr()),
             TextButton(
                 onPressed: () {
                   GOTO.popUT(context);
 
-                    BlocProvider.of<ConvertAndReadingCubit>(context)
-                        .readDocumentDataAndListeningOnStream();
-
+                  BlocProvider.of<ConvertAndReadingCubit>(context)
+                      .readDocumentDataAndListeningOnStream();
                 },
-                child: Text("try",style: TextStyle(color:Colors.green.shade700,fontSize: 15),).tr()),
+                child: Text(
+                  "try",
+                  style: TextStyle(color: Colors.green.shade700, fontSize: 15),
+                ).tr()),
           ],
         );
       });
 }
 
-connectionDialog(BuildContext context){
-  showDialog(context: context, builder: (c){
-    return AlertDialog(
-
-      title: Center(child: Text("noinet",style: TextStyle(fontSize: 16,color: Colors.grey.shade600,fontWeight: FontWeight.w600),).tr()),
-      actions:[ TextButton(
-          onPressed: () {
-            GOTO.pop(context);
-          },
-                child: Text("try",style: TextStyle(color:Colors.green.shade700,fontSize: 17),).tr()),
-          ],
-
-    );
-  });
-}
-
-deleteItemDialog(BuildContext context, AudioModel audioModel) {
+connectionDialog(BuildContext context) {
   showDialog(
       context: context,
       builder: (c) {
         return AlertDialog(
-          title:  Text(
+          title: Center(
+              child: Text(
+            "noinet",
+            style: TextStyle(
+                fontSize: 16,
+                color: Colors.grey.shade600,
+                fontWeight: FontWeight.w600),
+          ).tr()),
+          actions: [
+            TextButton(
+                onPressed: () {
+                  GOTO.pop(context);
+                },
+                child: Text(
+                  "try",
+                  style: TextStyle(color: Colors.green.shade700, fontSize: 17),
+                ).tr()),
+          ],
+        );
+      });
+}
+
+deleteItemDialog(BuildContext context, AudioModel audioModel, int index) {
+  showDialog(
+      context: context,
+      builder: (c) {
+        return AlertDialog(
+          title: Text(
             "confirm",
             style: TextStyle(
                 fontSize: 14,
@@ -150,7 +180,7 @@ deleteItemDialog(BuildContext context, AudioModel audioModel) {
                 color: Color.fromARGB(255, 49, 49, 49),
                 fontStyle: FontStyle.normal),
           ).tr(),
-          content:  Text(
+          content: Text(
             "sure",
             style: TextStyle(
                 fontSize: 14,
@@ -162,34 +192,43 @@ deleteItemDialog(BuildContext context, AudioModel audioModel) {
           actions: [
             TextButton(
                 onPressed: () async {
+                  print("----------" + audioModel.path);
                   File aufioFile = File(audioModel.path);
+                  bool hasFile = await aufioFile.exists();
+                  if (!hasFile) {
+                    GOTO.pop(context);
+                    return;
+                  }
                   await aufioFile.delete();
 
                   List<dynamic> listMap =
-                  await HiveDB.loadCountryCode(key: "listOfAudio");
+                      await HiveDB.loadCountryCode(key: "listOfAudio");
 
-                  List<AudioModel> listOfAudioModels = [];
-                  List<dynamic> listOfJson = [];
+                  AudioBook book = AudioBook.fromJson(listMap[index]);
 
-                  listOfAudioModels = List.generate(listMap.length,
-                          (index) => AudioModel.fromJson(listMap[index]));
+                  book.partOfBook!.remove(audioModel);
 
-                  listOfAudioModels.remove(audioModel);
+                  listMap.removeAt(index);
+                  listMap.insert(index, book.toJson());
 
-                  listOfJson = List.generate(listOfAudioModels.length,
-                          (index) => listOfAudioModels[index].toJson());
-
-                  await HiveDB.saveData("listOfAudio", listOfJson);
+                  await HiveDB.saveData("listOfAudio", listMap);
                   GOTO.pop(context);
 
-                  BlocProvider.of<SavedBooksCubit>(context).loadList();
+                  BlocProvider.of<PartAudioBooksDartCubit>(context)
+                      .loadList(index);
                 },
-                child:  Text("delete",style: TextStyle(color:Colors.green.shade700,fontSize: 15)).tr()),
+                child: Text("delete",
+                        style: TextStyle(
+                            color: Colors.green.shade700, fontSize: 15))
+                    .tr()),
             TextButton(
                 onPressed: () {
                   GOTO.pop(context);
                 },
-                child:  Text("no",style: TextStyle(color:Colors.green.shade700,fontSize: 15)).tr()),
+                child: Text("no",
+                        style: TextStyle(
+                            color: Colors.green.shade700, fontSize: 15))
+                    .tr()),
           ],
         );
       });
